requirements.txt
apparmor==2.13.3
appdirs==1.4.3
application-utility==1.3.1
asgiref==3.2.5
astroid==2.3.3
bauh==0.8.3
beautifulsoup4==4.8.2
btrfsutil==1.1.1
CacheControl==0.12.6
caffeine-ng==3.4.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
colorama==0.4.3
contextlib2==0.6.0
cupshelpers==1.0
Cython==0.29.15
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
docopt==0.6.2
ewmh==0.1.6
gestures==0.2.2
gufw==19.10.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
layoutswitcherlib==0.8.2
lazy-object-proxy==1.4.3
LibAppArmor==2.13.3
louis==3.13.0
mccabe==0.6.1
more-itertools==8.2.0
msgpack==0.6.2
npyscreen==4.10.5
ordered-set==3.1.1
packaging==20.2
pacman-mirrors==4.16.4
pep517==0.8.1
Pillow==6.2.1
progress==1.5
psycopg2==2.8.4
pwquality==1.4.2
pyaml==19.12.0
pycairo==1.19.1
pycups==1.9.74
pycurl==7.43.0.5
PyGObject==3.34.0
pylint==2.4.4
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.1
pysmbc==1.0.18
python-xlib==0.26
pytoml==0.1.21
pytz==2019.3
pyxdg==0.26
PyYAML==5.3
reportlab==3.5.34
requests==2.23.0
retrying==1.3.3
setproctitle==1.1.10
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.1
team==1.0
toml==0.10.0
ufw==0.36
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.1.0
